+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased] - 2026-02-28
 ### Added
+- Added "Michi" link to the Web Project Examples section in Home.jsx and translations.
 - Added "Ristorante" link to the Web Project Examples section in Home.jsx and translations.
 - Added "Estetica" link to the Web Project Examples section in Home.jsx and translations.
 - Added "Dr. Paws Clinic" link to the Web Project Examples section in Home.jsx and translations.